+++ spec/unit/shokkenki/consumer/rspec/rspec_configuration_spec.rb
@@ -114,4 +114,29 @@ describe 'RSpec configuration' do
       expect(hooks).to have_received(:after_suite)
     end
   end
+
+  context 'when the RSpec version supports backtrace inclusion patterns' do
+    let(:patterns) { double('patterns').as_null_object }
+
+    before do
+      allow(config).to receive(:backtrace_inclusion_patterns).and_return patterns
+      load_config
+    end
+
+    it 'includes shokkenki-provider lines in the backtrace' do
+      expect(patterns).to have_received(:<<).with(/shokkenki\-consumer/)
+    end
+  end
+
+  context "when the RSpec version doesn't support backtrace inclusion patterns" do
+    let(:patterns) { double('patterns').as_null_object }
+
+    before do
+      load_config
+    end
+
+    it "doesn't attempt to add any patterns for exclusion" do
+      expect(patterns).to_not have_received(anything)
+    end
+  end
 end